Renamed files to start with capital letters, fixed patient search bug where clearing caused the app to crash, and added functionality to search by date of birth and by phone number
