+++ client/src/App.tsx
@@ -8,9 +8,9 @@ import {
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./App.css";
 
-import Home from "./ehr_components/home";
+import Home from "./ehr_components/Home";
 import Main from "./ehr_components/Main";
-import Loader from "./components/Loader";
+import Loader from "./ehr_components/ui/Loader";
 import Auth from "./components/Auth";
 
 import { AuthContextProvider } from "./store/auth-context";
